Incomplete

+++ avs_dx/DxVisuals/Render/EffectsState.cpp
@@ -1,13 +1,13 @@
 #include "stdafx.h"
 #include "EffectsState.h"
 
-HRESULT EffectsState::create( int totalSize )
+HRESULT EffectsState::create( UINT totalSize )
 {
 	m_srv = nullptr;
 	m_uav = nullptr;
 
-	// "+1" for the frames counter.
-	const UINT byteWidth = ( (UINT)totalSize + 1 ) * 4;
+	// "+1" for the frame counter
+	const UINT byteWidth = ( totalSize + 1 ) * 4;
 	constexpr UINT bindFlags = D3D11_BIND_SHADER_RESOURCE | D3D11_BIND_UNORDERED_ACCESS;
 	CD3D11_BUFFER_DESC bufferDesc{ byteWidth, bindFlags, D3D11_USAGE_DEFAULT, 0, D3D11_RESOURCE_MISC_BUFFER_ALLOW_RAW_VIEWS };
